feat: Add behavior system.

+++ src/DG.Core/Components/Common/DGHealthComponent.cs
@@ -8,23 +8,23 @@
 
         public void SetCurrentHealth(int value)
         {
-            this.CurrentHealth = (int)value;
+            this.CurrentHealth = value;
         }
 
         public void SetMaximumHealth(int value)
         {
-            this.MaximumHealth = (int)value;
+            this.MaximumHealth = value;
         }
 
         public void Hurt(int value)
         {
-            this.CurrentHealth -= (int)value;
+            this.CurrentHealth -= value;
             this.CurrentHealth = this.CurrentHealth < 0 ? 0 : this.CurrentHealth;
         }
 
         public void Heal(int value)
         {
-            this.CurrentHealth += (int)value;
+            this.CurrentHealth += value;
             this.CurrentHealth = this.CurrentHealth > this.MaximumHealth ? this.MaximumHealth : this.CurrentHealth;
         }
     }
